# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Satara Development Internal Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List units
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  # ─── CONSTRUCTION ─────────────────────────────────────────────
...
  /qc/defects:
    get:
      operationId: listQcDefects
      tags: [qc]
      summary: List QC defects
      parameters:
        - name: unitId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of defects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcDefect"
    post:
      operationId: createQcDefect
      tags: [qc]
      summary: Report QC defect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcDefectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcDefect"
  /qc/defects/{id}:
    patch:
      operationId: updateQcDefect
      tags: [qc]
      summary: Update QC defect status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcDefectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcDefect"

  # ─── MATERIALS ───────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, nama, lokasi, fase, status, totalUnit, createdAt]
      properties:
        id: { type: integer }
        nama: { type: string }
        lokasi: { type: string }
        provinsi: { type: ["string", "null"] }
        kabupaten: { type: ["string", "null"] }
        kecamatan: { type: ["string", "null"] }
        desa: { type: ["string", "null"] }
        luas: { type: ["number", "null"] }
        totalUnit: { type: integer }
        fase:
          type: string
          enum: [LAND, PLAN, LEGAL, SELL, BUILD, AKAD, HANDOVER, SCALE]
        status:
          type: string
          enum: [active, completed, on_hold]
        targetStart: { type: ["string", "null"] }
        targetEnd: { type: ["string", "null"] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
        createdAt: { type: string }
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            units:
              type: array
              items:
                $ref: "#/components/schemas/Unit"
            landProspects:
              type: array
              items:
                $ref: "#/components/schemas/LandProspect"
            legalDocuments:
              type: array
              items:
                $ref: "#/components/schemas/LegalDocument"
...
    ProjectInput:
      type: object
      required: [nama, lokasi, fase, totalUnit]
      properties:
        nama: { type: string }
        lokasi: { type: string }
        provinsi: { type: string }
        kabupaten: { type: string }
        kecamatan: { type: string }
        desa: { type: string }
        luas: { type: number }
        totalUnit: { type: integer }
        fase:
          type: string
          enum: [LAND, PLAN, LEGAL, SELL, BUILD, AKAD, HANDOVER, SCALE]
        status:
          type: string
          enum: [active, completed, on_hold]
        targetStart: { type: string }
        targetEnd: { type: string }
        lat: { type: number }
        lng: { type: number }
    ProjectUpdate:
      type: object
      properties:
        nama: { type: string }
        lokasi: { type: string }
        provinsi: { type: string }
        kabupaten: { type: string }
        kecamatan: { type: string }
        desa: { type: string }
        luas: { type: number }
        totalUnit: { type: integer }
        fase:
          type: string
          enum: [LAND, PLAN, LEGAL, SELL, BUILD, AKAD, HANDOVER, SCALE]
        status:
          type: string
          enum: [active, completed, on_hold]
        targetStart: { type: string }
        targetEnd: { type: string }
        lat: { type: number }
        lng: { type: number }

    # ─── LAND PROSPECT ───────────────────────────────────────────
    LandProspect:
      type: object
      required: [id, lokasi, luas, hargaM2, status, roi, margin, projectId, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: ["integer", "null"] }
        lokasi: { type: string }
        luas: { type: number }
        hargaM2: { type: number }
        status:
          type: string
          enum: [prospek_baru, survey, analisis_kompetitor, negosiasi, legal_checking, pks_mou, ditolak]
        roi: { type: number }
        margin: { type: number }
        aksesJalan: { type: ["number", "null"] }
        riskLevel:
          type: ["string", "null"]
          enum: [green, yellow, red, null]
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
...
    LegalDocument:
      type: object
      required: [id, projectId, tipeDokumen, status, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        tipeDokumen:
          type: string
          enum: [SHM, AJB, balik_nama, split, PBB, PKKPR, SPPL, PBG, SLF, bank_ready, sikumbang, lainnya]
        status:
          type: string
          enum: [pending, in_progress, approved, rejected]
        pic: { type: ["string", "null"] }
        expiry: { type: ["string", "null"] }
        catatan: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Unit:
      type: object
      required: [id, projectId, blok, nomor, tipe, harga, status, progress, createdAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        customerId: { type: ["integer", "null"] }
        blok: { type: string }
        nomor: { type: string }
        tipe: { type: string }
        harga: { type: number }
        status:
          type: string
          enum: [available, booked, kpr_process, ready_akad, akad, serah_terima]
        progress: { type: number }
        readyAkad: { type: boolean }
        createdAt: { type: string }
    UnitInput:
      type: object
      required: [projectId, blok, nomor, tipe, harga]
      properties:
        projectId: { type: integer }
        blok: { type: string }
        nomor: { type: string }
        tipe: { type: string }
        harga: { type: number }
        status:
          type: string
          enum: [available, booked, kpr_process, ready_akad, akad, serah_terima]
    UnitUpdate:
      type: object
      properties:
        customerId: { type: integer }
        status:
          type: string
          enum: [available, booked, kpr_process, ready_akad, akad, serah_terima]
        readyAkad: { type: boolean }

    # ─── CONSTRUCTION ─────────────────────────────────────────────
...
    QcDefect:
      type: object
      required: [id, unitId, kategori, deskripsi, status, createdAt]
      properties:
        id: { type: integer }
        unitId: { type: integer }
        kategori: { type: string }
        deskripsi: { type: string }
        status:
          type: string
          enum: [open, in_repair, verified, closed]
        verifiedBy: { type: ["string", "null"] }
        createdAt: { type: string }
    QcDefectInput:
      type: object
      required: [unitId, kategori, deskripsi]
      properties:
        unitId: { type: integer }
        kategori: { type: string }
        deskripsi: { type: string }
    QcDefectUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_repair, verified, closed]
        verifiedBy: { type: string }

    # ─── MATERIAL ─────────────────────────────────────────────────
...