# lib/api-zod/src/generated/types/unitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Satara Development Internal Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { UnitInputStatus } from './unitInputStatus';

export interface UnitInput {
  projectId: number;
  blok: string;
  nomor: string;
  tipe: string;
  harga: number;
  status?: UnitInputStatus;
}