# lib/api-zod/src/generated/types/unitUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Satara Development Internal Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { UnitUpdateStatus } from './unitUpdateStatus';

export interface UnitUpdate {
  customerId?: number;
  status?: UnitUpdateStatus;
  readyAkad?: boolean;
}